+++ app/Meeting.php
@@ -30,12 +30,19 @@ use Illuminate\Database\Eloquent\Model;
 class Meeting extends Model
 {
     /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table = 'meetings';
+
+    /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
     protected $fillable = ['title', 'description', 'start_time',
-                    'end_time', 'repeat', 'repetition_end_time'];
+                    'end_time', 'repeat', 'repetition_end_time', 'group_id'];
 
     /**
      * @return \Illuminate\Database\Eloquent\Relations\BelongsToMany
